moved PricingService::getMinimumPriceByPriceWithoutVat to ProductPriceCalculation::getMinimumPriceByPriceWithoutVat
