+++ packages/framework/src/Model/Product/Pricing/ProductPriceCalculation.php
@@ -95,7 +95,7 @@ class ProductPriceCalculation
             $variantPrices[] = $this->calculatePrice($variant, $domainId, $pricingGroup);
         }
 
-        $minVariantPrice = $this->pricingService->getMinimumPriceByPriceWithoutVat($variantPrices);
+        $minVariantPrice = $this->getMinimumPriceByPriceWithoutVat($variantPrices);
         $from = $this->pricingService->arePricesDifferent($variantPrices);
 
         return new ProductPrice($minVariantPrice, $from);
@@ -122,5 +122,25 @@ class ProductPriceCalculation
         );
 
         return new ProductPrice($basePrice, false);
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Pricing\Price[] $prices
+     * @return \Shopsys\FrameworkBundle\Model\Pricing\Price
+     */
+    public function getMinimumPriceByPriceWithoutVat(array $prices)
+    {
+        if (count($prices) === 0) {
+            throw new \Shopsys\FrameworkBundle\Model\Pricing\Exception\InvalidArgumentException('Array can not be empty.');
+        }
+
+        $minimumPrice = null;
+        foreach ($prices as $price) {
+            if ($minimumPrice === null || $price->getPriceWithoutVat() < $minimumPrice->getPriceWithoutVat()) {
+                $minimumPrice = $price;
+            }
+        }
+
+        return $minimumPrice;
     }
 }
